+++ src/Utils/MIME.jl
@@ -19,25 +19,30 @@ get_mime_type("image.png")  # Returns "image/png"
 ```
 """
 function get_mime_type(filename::String)
-    ext = lowercase(split(filename, '.')[end])
+    dot_idx = findlast(==('.'), filename)
+    if dot_idx === nothing
+        return "application/octet-stream"
+    end
     
-    mime_types = Dict(
-        "html" => "text/html",
-        "css" => "text/css",
-        "js" => "application/javascript",
-        "svg" => "image/svg+xml",
-        "png" => "image/png",
-        "jpg" => "image/jpeg",
-        "jpeg" => "image/jpeg",
-        "gif" => "image/gif",
-        "ico" => "image/x-icon",
-        "json" => "application/json",
-        "txt" => "text/plain",
-        "woff" => "font/woff",
-        "woff2" => "font/woff2",
-        "ttf" => "font/ttf",
-        "otf" => "font/otf"
-    )
+    ext = lowercase(SubString(filename, dot_idx + 1))
     
-    return get(mime_types, ext, "application/octet-stream")
+    return get(MIME_TYPES, ext, "application/octet-stream")
 end
+
+const MIME_TYPES = Dict(
+    "html" => "text/html",
+    "css" => "text/css",
+    "js" => "application/javascript",
+    "svg" => "image/svg+xml",
+    "png" => "image/png",
+    "jpg" => "image/jpeg",
+    "jpeg" => "image/jpeg",
+    "gif" => "image/gif",
+    "ico" => "image/x-icon",
+    "json" => "application/json",
+    "txt" => "text/plain",
+    "woff" => "font/woff",
+    "woff2" => "font/woff2",
+    "ttf" => "font/ttf",
+    "otf" => "font/otf"
+)